Refactor [all-actions] src/*
 - try to move all the `all actions` into a separate folder for better tree-shaking
