+++ src/index.ts
@@ -3,8 +3,9 @@ export * from './collections';
 export * from './meta-events';
 export * from './subscribe';
 export * from './events';
+export * from './debug';
+export * from './types';
 export * from './once';
 export * from './emit';
 export * from './wait';
-export * from './debug';
-export * from './types';
+export * from './all';
