+++ src/util.ts
@@ -1,10 +1,4 @@
-import { TEventMap, THandlerOf } from './types';
-
-export type TDoAction<P extends any[] = any[], R = void> = <M extends TEventMap>(
-  eventMap: M
-) => <E extends keyof M>(
-  event: E
-) => (...args: P) => R;
+import { TEventMap } from './types';
 
 /**
  * Maps object values by their keys into a new object
@@ -21,26 +15,8 @@ export const mapObject = <T extends object, R>(
   newObj[key] = value(key), newObj
 ), <Record<keyof T, R>> {});
 
-/**
- * A `do`-er factory
- *
- * Applies a specified action for all events in a collection
- *
- * @param action - an action to apply
- */
-export const doForAll = <A extends TDoAction>(
-  action: A
-) => <M extends TEventMap>(
+export type TDoAction<P extends any[] = any[], R = void> = <M extends TEventMap>(
   eventMap: M
-) => (
-  ...args: A extends TDoAction<infer P> ? P : any[]
-) => {
-  mapObject(eventMap, key => action(eventMap)(key)(...args));
-};
-
-export type THandlersForAll = {
-  <M extends TEventMap>(eventMap: M): {
-    (handler: THandlerOf<M>): void;
-    (...handlers: THandlerOf<M>[]): void;
-  };
-};
+) => <E extends keyof M>(
+  event: E
+) => (...args: P) => R;
